+++ CHANGELOG.md
@@ -4,11 +4,45 @@
 
 ### Features
 
-- ASAB WebUI submodule version bump [[98343e2](https://github.com/TeskaLabs/asab-webui/commit/https://github.com/TeskaLabs/asab-webui/commit/98343e2ce799fc923d8ec2dd5eb67b19e2df5d93)] commit (INDIGO Sprint 221209, [!19](https://github.com/TeskaLabs/seacat-auth-webui/pull/19))
+- ASAB WebUI submodule version bump [[371d35c](https://github.com/TeskaLabs/asab-webui/commit/371d35cd7737f4a3484499ff37d84d067787db87)] commit (INDIGO Sprint 230331, [!26](https://github.com/TeskaLabs/seacat-auth-webui/pull/26))
+
+### Refactoring
+
+- Remove `external_code` from localstorage when login in with external services, update eternal login error messages, polish code (INDIGO Sprint 230331, [!26](https://github.com/TeskaLabs/seacat-auth-webui/pull/26))
+
+## v23.13-beta
+
+### Features
+
+- ASAB WebUI submodule version bump [[1960466](https://github.com/TeskaLabs/asab-webui/commit/1960466a45d0c48ec279e703317ebf0a59fdcbda)] commit (INDIGO Sprint 230317, [!25](https://github.com/TeskaLabs/seacat-auth-webui/pull/25))
+
+### Refactoring
+
+- Remove react-helmet from the package.json (INDIGO Sprint 230217, [!22](https://github.com/TeskaLabs/seacat-auth-webui/pull/22))
+
+- Webpack V4 to V5 migration. (INDIGO Sprint 230303, [!17](https://github.com/TeskaLabs/seacat-auth-webui/pull/17))
+
+
+## v23.4
+
+### Compatibility
+
+- [x] **SeaCat Admin UI `v23.4` or newer**
+- [x] **SeaCat Auth `v23.3` or newer**
+
+⚠️ Due to implementation of new features, invitation will not work with SeaCat Auth service versions older than `v23.3`
+
+### Features
+
+- ASAB WebUI submodule version bump [[25d5a5c](https://github.com/TeskaLabs/asab-webui/commit/https://github.com/TeskaLabs/asab-webui/commit/25d5a5ce97c6f2286525127cc3a31531b03312f3)] commit (INDIGO Sprint 230106, [!16](https://github.com/TeskaLabs/seacat-auth-webui/pull/16))
+
+- Implement registration by invitation - create invitation and registration screens and cards (INDIGO Sprint 230106, [!16](https://github.com/TeskaLabs/seacat-auth-webui/pull/16))
 
 ### Refactoring
 
 - Replace `phone_number` with `phone` and `preferred_username` with `username` (INDIGO Sprint 221209, [!19](https://github.com/TeskaLabs/seacat-auth-webui/pull/19))
+
+- Update alert responses with full message from the service and prolong the time of the error message to 30s (INDIGO Sprint 230106, [!16](https://github.com/TeskaLabs/seacat-auth-webui/pull/16))
 
 ## v22.48